Refactor to separate nodes from programs. This allows us to unit test the nodes.
